+++ CHANGES.rst
@@ -37,9 +37,10 @@ Changes
 - The `packaging` dependency was removed.
   :pr:`1307` by :user:`Jovan Stojanovic <jovan-stojanovic>`
 
-- :class:`TextEncoder`, :class:`StringEncoder` and :class:`GapEncoder` outputs during
-  transform are now normalized by computing the total standard deviation during
-  training. :pr:`1274` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
+- :class:`TextEncoder`, :class:`StringEncoder` and :class:`GapEncoder` now compute the
+  total standard deviation norm during training, which is a global constant, and
+  normalize the vector outputs by performing element-wise division on all entries.
+  :pr:`1274` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
 
 Bugfixes
 --------